refactor(docs): progress dx
