+++ apps/docs/content/components/progress/indeterminate.raw.jsx
@@ -1,0 +1,5 @@
+import {Progress} from "@nextui-org/react";
+
+export default function App() {
+  return <Progress isIndeterminate aria-label="Loading..." className="max-w-md" size="sm" />;
+}
