+++ apps/docs/content/components/progress/indeterminate.ts
@@ -1,15 +1,4 @@
-const App = `import { Progress } from "@nextui-org/react";
-
-export default function App() {
-  return (
-    <Progress
-      size="sm"
-      isIndeterminate
-      aria-label="Loading..."
-      className="max-w-md"
-    />
-  );
-}`;
+import App from "./indeterminate.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
