+++ apps/docs/content/components/progress/label.raw.jsx
@@ -1,0 +1,5 @@
+import {Progress} from "@nextui-org/react";
+
+export default function App() {
+  return <Progress className="max-w-md" label="Loading..." value={55} />;
+}
