+++ apps/docs/content/components/progress/label.ts
@@ -1,10 +1,4 @@
-const App = `import {Progress} from "@nextui-org/react";
-
-export default function App() {
-  return (
-    <Progress label="Loading..." value={55}  className="max-w-md" />
-  );
-}`;
+import App from "./label.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
