+++ apps/docs/content/components/progress/striped.raw.jsx
@@ -1,0 +1,7 @@
+import {Progress} from "@nextui-org/react";
+
+export default function App() {
+  return (
+    <Progress isStriped aria-label="Loading..." className="max-w-md" color="secondary" value={60} />
+  );
+}
