+++ apps/docs/content/components/progress/striped.ts
@@ -1,17 +1,4 @@
-const App = `import { Progress } from "@nextui-org/react";
-
-export default function App() {
-  return (
-    <Progress
-      isStriped
-      aria-label="Loading..."
-      color="secondary"
-      value={60}
-      className="max-w-md"
-    />
-  );
-}
-`;
+import App from "./striped.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
