+++ apps/docs/content/components/progress/usage.raw.jsx
@@ -1,0 +1,5 @@
+import {Progress} from "@nextui-org/react";
+
+export default function App() {
+  return <Progress aria-label="Loading..." className="max-w-md" value={60} />;
+}
