+++ apps/docs/content/components/progress/usage.ts
@@ -1,10 +1,4 @@
-const App = `import {Progress} from "@nextui-org/react";
-
-export default function App() {
-  return (
-    <Progress aria-label="Loading..." value={60} className="max-w-md"/>
-  );
-}`;
+import App from "./usage.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
