+++ apps/docs/content/components/progress/value-formatting.ts
@@ -1,19 +1,4 @@
-const App = `import {Progress} from "@nextui-org/react";
-
-export default function App() {
-  return (
-    <Progress
-      label="Monthly expenses"
-      size="sm"
-      value={4000}
-      maxValue={10000}
-      color="warning"
-      formatOptions={{style: "currency", currency: "ARS"}}
-      showValueLabel={true}
-      className="max-w-md"
-    />
-  );
-}`;
+import App from "./value-formatting.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
